First pass draft
Next steps are:

1 - to redo `_eT_zhz_einsum_vibrational_components` and all its sub functions.
Right now the logic is setup for [h, t, t, t ...] but we want a
[t, t, t, ... h, z] structure so we need to change how the indexing works

2 - redo `_eT_zhz_einsum_electronic_components`

an example:
R += (
    np.einsum('ac, cdi, db -> ab', t_args[(0, 0)], h_args[(0, 1)], z_args[(2, 0)]) +
    np.einsum('ac, cdz, db -> abz', t_args[(0, 0)], h_args[(1, 0)], z_args[(0, 0)]) +
    np.einsum('ac, cdiz, db -> abz', t_args[(0, 0)], h_args[(1, 1)], z_args[(1, 0)])
)

here we can see that the first term results in `ab` instead of `abz`
because the `z` index is not present on the left hand side, because
we are missing the logic to add the `z` to the `db` which represents
the z_right term. The right hand side is always `ab` plus all non traced
terms.

3 - need to look at the end of the function `_generate_eT_zhz_einsums`
when we call `_write_third_eTz_einsum_python` and pass in the list of terms
we do first a 1 * H * Z, that is the `valid_zero_list`
and then we do a e^T * H * Z, that is the `valid_term_list`

_write_third_eTz_einsum_python(LHS.rank, operators, valid_zero_list)
_write_third_eTz_einsum_python(LHS.rank, operators, valid_term_list)

for the `valid_zero_list` terms .... I THINK we just don't want the
`t_args[(0, 0)]` contributions, however this bears more investigation