Temp fix - module logging config from driver.py
This solution is not desirable but it is functional for now and we can
move onto other stuff and improve it later.

This allows the user to specify the logging path on the command line and
all logging objects in the library respect the user provided argument,
while also being global objects within their own modules.